% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.convergence.R
\name{check.convergence}
\alias{check.convergence}
\title{Checks whether the model has converged}
\usage{
check.convergence(ddf.model)
}
\arguments{
\item{ddf.model}{ddf object}
}
\value{
boolean
}
\description{
Checks whether the model has converged
}
\author{
Laura Marshall
}
