% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check.fitted.R
\name{check.fitted}
\alias{check.fitted}
\title{Checks whether the model's fitted values make sense}
\usage{
check.fitted(ddf.model)
}
\arguments{
\item{ddf.model}{ddf object}
}
\value{
boolean
}
\description{
Checks whether the model's fitted values make sense
}
\author{
Laura Marshall
}

