% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mae.warning}
\alias{mae.warning}
\title{Warning function}
\usage{
mae.warning(warning.msg = NULL, warning.mode = "store", MAE.warnings)
}
\arguments{
\item{warning.msg}{the message to be stored/printed (optional)}

\item{warning.mode}{report or print errors (default report)}

\item{MAE.warnings}{character vector of existing warning messages}
}
\value{
None
}
\description{
Writes or stores messages for various situations that can occur
}
\author{
Dave Miller & Laura Marshall
}

