\name{process.warnings}
\alias{process.warnings}
\title{Summarises warnings}
\usage{
  process.warnings(MAE.warnings)
}
\description{
  Sumarises warnings generated during the bootstrap and
  removes the MAE.warnings global object.
}
\note{
  Internal function not intended to be called by user.
}
\author{
  Laura Marshall
}

