% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.param.ests.R
\name{store.param.ests}
\alias{store.param.ests}
\title{Updates bootstrap.ddf.statistics}
\usage{
store.param.ests(
  bootstrap.ddf.statistics,
  species.name,
  model.name,
  ddf.model,
  rep.no
)
}
\arguments{
\item{bootstrap.ddf.statistics}{a list containing various arrays and vectors
storing model details.}

\item{species.name}{name of species the model relates to}

\item{model.name}{name of ddf object}

\item{ddf.model}{ddf object}

\item{rep.no}{iteration number of the bootstrap}
}
\value{
bootstrap.ddf.statistics the updated list
}
\description{
#' Updates bootstrap.ddf.statistics with the latest parameter estimates from
  the ddf object supplied.
}
\note{
Internal function not intended to be called by user.
}
\seealso{
\code{create.param.arrays}
}
\author{
Laura Marshall
}
\keyword{internal}
