% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ma.analysis.R
\name{summary.ma.analysis}
\alias{summary.ma.analysis}
\title{Summary of multi-analysis object}
\usage{
\method{summary}{ma.analysis}(object, ...)
}
\arguments{
\item{object}{a \code{ma} model object}

\item{\dots}{unspecified and unused arguments for S3 consistency}
}
\value{
list of extracted and summarized objects
}
\description{
Provides a summary of the fitted detection probability model
parameters, model selection criterion, and optionally abundance in the
covered (sampled) region and its standard error for all species.
}
\note{
This function is called by the generic function \code{summary} for any
  \code{ma} object.
}
\author{
Laura Marshall
}
\keyword{utility}
