% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.ma}
\alias{summary.ma}
\title{Summary of multi-analysis object}
\usage{
\method{summary}{ma}(object, description = FALSE, glossary = FALSE, ...)
}
\arguments{
\item{object}{an object of class \code{ma}}

\item{description}{boolean if you would like}

\item{glossary}{a \code{ma} model object would like a glossary to
be displayed}

\item{\dots}{unspecified and unused arguments for S3 consistency}
}
\value{
list of extracted and summarized objects
}
\description{
Provides asummary of the fitted detection probability model
parameters, model selection criterion, and optionally abundance in the
covered (sampled) region and its standard error for all species.
}
\note{
This function is called by the generic function \code{summary} for any
  \code{ma} object.
}
\author{
Laura Marshall
}
\keyword{utility}

