% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.ma.species.R
\name{summary.ma.species}
\alias{summary.ma.species}
\title{Print a summary of an element of a multi-analysis result correcponding to
a single species included in the analyses.}
\usage{
\method{summary}{ma.species}(object, species = NULL, ...)
}
\arguments{
\item{object}{a summary of \code{ma} model object}

\item{species}{optional character value giving the species name, solely for
display purposes}

\item{\dots}{unspecified and unused arguments for S3 consistency}
}
\description{
Provides a summary of the fitted detection probability model
parameters, model selection criterion, and optionally abundance in the
covered (sampled) region and its standard error. What is printed depends
on the corresponding call to summary.
}
\author{
Laura Marshall
}
\seealso{
\code{\link{summary.ma}}
}
\keyword{utility}

