% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.ma.unid.R
\name{summary.ma.unid}
\alias{summary.ma.unid}
\title{Summary of multi-analysis object}
\usage{
\method{summary}{ma.unid}(object, species = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{ma.unid}}

\item{species}{optional character value giving the species name, solely for
display purposes}

\item{\dots}{unspecified and unused arguments for S3 consistency}
}
\value{
list of extracted and summarized objects
}
\description{
Provides asummary of the fitted detection probability model
parameters, model selection criterion, and optionally abundance in the
covered (sampled) region and its standard error for all species.
}
\note{
This function is called by the generic function \code{summary} for any
  \code{ma} object.
}
\author{
Laura Marshall
}
\keyword{utility}

