% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumulate.results.R
\name{accumulate.results}
\alias{accumulate.results}
\title{Enters the prorated results into the bootstrap.results array}
\usage{
accumulate.results(n, bootstrap.results, formatted.results, clusters)
}
\arguments{
\item{n}{index of the current bootstrap iteration}

\item{bootstrap.results}{list of 4-dimensional arrays containing the 
bootstrap results}

\item{formatted.results}{list of data objects similar to the dht class}

\item{clusters}{boolean are the observations clusters of individuals
bootstrap results}
}
\value{
list of 4-dimensional arrays containing the updated
}
\description{
Enters the prorated results into the bootstrap.results array
}
\author{
Laura Marshall
}
