% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check.ddf.models.R
\name{check.ddf.models}
\alias{check.ddf.models}
\title{Checks the list of model names supplied by the user}
\usage{
check.ddf.models(model.names, ddf.models)
}
\arguments{
\item{model.names}{a list of vectors of model names. Each list element is
named according to the species code.}

\item{ddf.models}{a list of ddf objects named in model.names}
}
\value{
list of two boolean values relating to cluster size and double
  observer analyses.
}
\description{
Performs various checks on the models and model names supplied by the user,
including checking that the data in each model are the same across species
codes to ensure valid comparison using the AIC/AICc/BIC selection criteria.
}
\details{
In addition, this routine establishes whether the analyses are double
observer or not and whether the observations are individuals of clusters.
}
\note{
Internal function not intended to be called by user.
}
\author{
Laura Marshall
}
\seealso{
\code{execute.multi.analysis}
}
\keyword{data}
\keyword{input}
\keyword{validation}
\keyword{validation,}

