% Generated by roxygen2 (4.0.2): do not edit by hand
\name{process.warnings}
\alias{process.warnings}
\title{Summarises warnings}
\usage{
process.warnings(MAE.warnings)
}
\arguments{
\item{MAE.warnings}{character vector of warning messages}
}
\description{
Sumarises warnings generated during the bootstrap and removes the
MAE.warnings global object.
}
\note{
Internal function not intended to be called by user.
}
\author{
Laura Marshall
}

