% Generated by roxygen2 (4.0.2): do not edit by hand
\name{resample.covariates}
\alias{resample.covariates}
\title{Parametrcally resamples selected variables in the dataset.}
\usage{
resample.covariates(ddf.dat.working, covariate.uncertainty, MAE.warnings)
}
\arguments{
\item{ddf.dat.working}{a list of datasets to be used in the analyses}

\item{covariate.uncertainty}{a dataframe detailing the variables to be
resampled - variable.layer, variable.name, cor.factor.layer,
cor.factor.name , uncertainty.layer, uncertainty.name,
uncertainty.measure, sampling.distribution.}

\item{MAE.warnings}{character vector of warning messages}
}
\value{
list of dataframes containing the parametrically resampled data
}
\description{
Parametricelly resamples specified variables in the dataset based on a
measure of uncertainty and a sampling distribution supplied by the user.
}
\details{
This function may also apply a correction factor. If numeric is specified the
same value will be applied to all observations. If a data layer is specified
then this function looks for a variable name in the specified layer of the
dataset and applies this as a correction factor.
}
\note{
Internal function not intended to be called by user.
}
\author{
Laura Marshall
}
\keyword{data}
\keyword{manipulation}

