% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtpois.R
\name{rtpois}
\alias{rtpois}
\title{Randomly generates values from a zero-truncated Poisson distribution}
\usage{
rtpois(N, mean = NA)
}
\arguments{
\item{N}{number of values to randomly generate}

\item{mean}{mean of the generated values}
}
\description{
Generates values from a zero-truncated Poisson distribution with mean
equal to that specified. It uses a look up table to check which value of 
lambda will give values with the requested mean.
}
\note{
Internal function not intended to be called by user.
}
\author{
Laura Marshall
}
